fix(exfil): use byte reads for live status; fix(radar): hysteresis buffering for map layer transitions
Exfil.cs
- Narrow Update() parameter from int to byte. Game memory stores exfil
  status as a single byte; reading as int was wider than the actual type.

ExfilManager.cs
- Halve refresh interval 3s → 1s for more responsive exfil status updates.
- AddRound(false): disable scatter-read cache so status reads always
  return live memory values, not a stale cached frame.
- Change scatter read/result type int → byte to match Exfil.Update(byte).

RadarMap.cs
- Track active layer in persistent _activeLayerIndex field (init = -1).
- Add 0.75 m hysteresis buffer on layer transitions to eliminate rapid
  floor-boundary flipping. Up transitions require playerHeight to exceed
  currentLayer.MaxHeight + buffer; down transitions require playerHeight
  to drop below currentLayer.MinHeight - buffer. Handles layers with
  missing Min/MaxHeight by falling back to the candidate layer's bound.
- Fix Pass 2 drawing predicate: always draw the active layer, the base
  layer, and any intermediate layers whose height range matches, instead
  of skipping non-matching layers unconditionally.

+++ src-silk/Tarkov/GameWorld/Exits/Exfil.cs
@@ -80,9 +80,7 @@ namespace eft_dma_radar.Silk.Tarkov.GameWorld.Exits
         }
 
         /// <summary>
-        /// Updates status and reads eligible entry points / scav IDs.
-        /// </summary>
-        public void Update(int rawStatus)
+        public void Update(byte rawStatus)
         {
             Status = rawStatus switch
             {
